requirements.txt
tensorflow
ultralytics
streamlit
matplotlib
seaborn
scikit-learn
Pillow
numpy
opencv-python
